projection with manual drop fields added
